+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # CHANGELOG
 
 ## 1.1.5 - Work in progress
+ - Enh #184: Add `last-login-ip` capture capability (kartik-v)
  - Enh: Changed `View::render()` calls in views to use absolute paths (ajmedway)
  - Fix #169: Fix bug in ReCaptchaComponent (BuTaMuH)
  - Fix #168: Fix spelling in russian language (EvgenyOrekhov)